+++ src/components/Creater/AnimatedBanner.tsx
