+++ src/components/UI/MainPage/HeroSection.tsx
