+++ src/components/UI/MainPage/StatsSection.tsx
